NOJIRA-Move-projectconfig-to-call-manager
Move projectconfig package from bin-common-handler to bin-call-manager since
bin-call-manager is the only consumer. Add 3+ services admission rule for
bin-common-handler to prevent single-consumer packages in the shared library.

- bin-call-manager: Add pkg/projectconfig (moved from common-handler)
- bin-call-manager: Update imports in models/common/domain.go
- bin-call-manager: Update imports in pkg/recordinghandler/main.go
- bin-common-handler: Remove pkg/projectconfig package
- docs: Add common-handler admission rule to root CLAUDE.md
- docs: Add design document for the move
