+++ bin-call-manager/pkg/projectconfig/main.go
