+++ bin-call-manager/pkg/projectconfig/main_test.go
